+++ src/main/php/web/Request.class.php
@@ -125,8 +125,8 @@ class Request implements Value {
     // Merge parameters from URL and urlencoded payload.
     $query= $this->uri->query(false) ?? '';
     $type= Headers::parameterized()->parse($this->header('Content-Type', ''));
-    if ('application/x-www-form-urlencoded' === $type->value()) {
-      $data= Streams::readAll($this->input->incoming());
+    if ('application/x-www-form-urlencoded' === $type->value() && $stream= $this->input->incoming()) {
+      $data= Streams::readAll($stream);
       $this->stream= new MemoryInputStream($data);
       $query.= '&'.$data;
     }
@@ -255,7 +255,17 @@ class Request implements Value {
    * @return iterable
    */
   public function dispatch($path, $params= []) {
-    yield 'dispatch' => $this->uri()->using()->path($path)->params($params)->create();
+    if (false === ($p= strpos($path, '?'))) {
+      $uri= $this->uri()->using()->path($path)->params($params);
+    } else {
+      $uri= $this->uri()->using()->path(substr($path, 0, $p))->query(substr($path, $p + 1), false);
+
+      // Merge parameters instead of overwriting them all via `params()`
+      foreach ($params as $name => $value) {
+        $uri->param($name, $value);
+      }
+    }
+    yield 'dispatch' => $uri->create();
   }
 
   /**