+++ src/main/php/web/Response.class.php
@@ -121,16 +121,15 @@ class Response {
   /**
    * Flushes response
    *
-   * @param  web.io.Output $output
    * @return void
    * @throws lang.IllegalStateException
    */
-  public function flush($output= null) {
+  public function flush() {
     if ($this->flushed) {
       throw new IllegalStateException('Response already flushed');
     }
 
-    $this->begin($output ?: $this->output);
+    $this->begin($this->output);
   }
 
   /**
@@ -157,8 +156,13 @@ class Response {
    *
    * @param  int $size If omitted, uses chunked transfer encoding
    * @return io.streams.OutputStream
+   * @throws lang.IllegalStateException
    */
   public function stream($size= null) {
+    if ($this->flushed) {
+      throw new IllegalStateException('Response already flushed');
+    }
+
     if (null === $size) {
       $output= $this->output->stream();
     } else {
@@ -166,7 +170,7 @@ class Response {
       $output= $this->output;
     }
 
-    $this->flush($output);
+    $this->begin($output);
     return $output;
   }