+++ src/main/php/xp/web/dev/Console.class.php
@@ -1,6 +1,8 @@
 <?php namespace xp\web\dev;
 
-use web\{Filter, Response};
+use Closure, Throwable;
+use lang\XPException;
+use web\{Filter, Response, Error};
 
 /**
  * The development console captures content written via `var_dump()`,
@@ -8,7 +10,8 @@ use web\{Filter, Response};
  * inside an easily readable format above the real output, using a 200
  * HTTP response status.
  *
- * @see   php://ob_start
+ * @see   https://www.php.net/ob_start
+ * @test  web.unittest.server.ConsoleTest
  */
 class Console implements Filter {
   private $template;
@@ -19,20 +22,20 @@ class Console implements Filter {
   }
 
   /**
-   * Creates HTML table rows
+   * Transforms template
    *
-   * @param  [:var] $headers
-   * @return string
+   * @param  string $template
+   * @param  [:var] $context
    */
-  private function rows($headers) {
-    $r= '';
-    foreach ($headers as $name => $value) {
-      $r.= '<tr>
-        <td class="name">'.htmlspecialchars($name).'</td>
-        <td class="value">'.htmlspecialchars(implode(', ', $value)).'</td>
-      </tr>';
-    }
-    return $r;
+  private function transform($template, $context) {
+    return preg_replace_callback(
+      '/\{\{([^ }]+) ?([^}]+)?\}\}/',
+      function($m) use($context) {
+        $value= $context[$m[1]] ?? '';
+        return $value instanceof Closure ? $value($context[$m[2]] ?? '') : htmlspecialchars($value);
+      },
+      $template
+    );
   }
 
   /**
@@ -46,27 +49,49 @@ class Console implements Filter {
   public function filter($req, $res, $invocation) {
     $buffer= new Response(new Buffer());
 
+    $proceed= [];
     try {
       ob_start();
-      yield from $invocation->proceed($req, $buffer);
+      foreach ($invocation->proceed($req, $buffer) as $event => $arg) {
+        $proceed[$event]= $arg;
+      }
+      $status= 200;
+      $kind= 'debug';
+      $debug= ob_get_clean();
+    } catch (Throwable $t) {
+      $buffer->answer($status= $t instanceof Error ? $t->status() : 500);
+      $kind= 'error';
+      $debug= ob_get_clean()."\n".XPException::wrap($t)->toString();
     } finally {
       $buffer->end();
-      $debug= ob_get_clean();
+      yield from $proceed;
     }
 
+    $res->trace= $buffer->trace;
     $out= $buffer->output();
-    if (empty($debug)) {
+    if (0 === strlen($debug)) {
       $out->drain($res);
     } else {
-      $res->status(200, 'Debug');
-      $res->send(sprintf(
-        typeof($this)->getClassLoader()->getResource($this->template),
-        htmlspecialchars($debug),
-        $out->status,
-        htmlspecialchars($out->message),
-        $this->rows($out->headers),
-        htmlspecialchars($out->bytes)
-      ));
+      $res->trace+= ['console' => $kind];
+      $res->answer($status, $kind);
+      $res->send($this->transform(typeof($this)->getClassLoader()->getResource($this->template), [
+        'kind'     => $kind,
+        'debug'    => $debug,
+        'status'   => $out->status,
+        'message'  => $out->message,
+        'headers'  => $out->headers,
+        'contents' => $out->bytes,
+        '#rows'    => function($headers) {
+          $r= '';
+          foreach ($headers as $name => $value) {
+            $r.= '<tr>
+              <td class="name">'.htmlspecialchars($name).'</td>
+              <td class="value">'.htmlspecialchars(implode(', ', $value)).'</td>
+            </tr>';
+          }
+          return $r;
+        }
+      ]), 'text/html; charset='.\xp::ENCODING);
     }
   }
 }