+++ src/test/php/web/unittest/HeadersTest.class.php
@@ -45,12 +45,16 @@ class HeadersTest {
   #[Test]
   public function accept() {
     Assert::equals(
-      [
-        new Parameterized('text/html', []),
-        new Parameterized('application/json', ['q' => '0.9']),
-        new Parameterized('*/*', ['q' => '0.8']),
-      ],
-      Headers::values(Headers::parameterized())->parse('text/html, application/json;q=0.9, */*;q=0.8')
+      ['text/html' => 1.0, 'application/json' => 0.9, '*/*' => 0.8],
+      Headers::qfactors()->parse('text/html, application/json;q=0.9, */*;q=0.8')
+    );
+  }
+
+  #[Test]
+  public function accept_encoding() {
+    Assert::equals(
+      ['deflate' => 1.0, 'gzip' => 1.0, '*' => 0.5],
+      Headers::qfactors()->parse('deflate, gzip;q=1.0, *;q=0.5')
     );
   }